BAARC-119: Move nats client package to internal
(cherry picked from commit 96fdd4d385d0de9db81196e28e015099b0c71034)
