+++ cmd/arc/services/metamorph.go
@@ -13,7 +13,7 @@ import (
 
 	cfg "github.com/bitcoin-sv/arc/internal/helpers"
 	"github.com/bitcoin-sv/arc/internal/metamorph"
-	"github.com/bitcoin-sv/arc/internal/metamorph/async/nats_mq"
+	"github.com/bitcoin-sv/arc/internal/metamorph/async"
 	"github.com/bitcoin-sv/arc/internal/metamorph/store"
 	"github.com/bitcoin-sv/arc/internal/metamorph/store/postgresql"
 	"github.com/bitcoin-sv/arc/internal/version"
@@ -88,7 +88,7 @@ func StartMetamorph(logger *slog.Logger) (func(), error) {
 	capacityRequired := int(float64(targetTps*avgMinPerBlock*secPerMin) / float64(maxBatchSize))
 	minedTxsChan := make(chan *blocktx_api.TransactionBlocks, capacityRequired)
 
-	mqClient, err := nats_mq.NewNatsMQClient(minedTxsChan, logger, natsURL)
+	mqClient, err := async.NewNatsMQClient(minedTxsChan, logger, natsURL)
 	if err != nil {
 		return nil, err
 	}
